refactor: move builderIdStartUrl to constants file
