+++ packages/core/src/auth/sso/constants.ts
@@ -8,6 +8,8 @@
  * web, node, or vue.
  */
 
+export const builderIdStartUrl = 'https://view.awsapps.com/start'
+
 export const ssoUrlFormatRegex =
     /^(https?:\/\/(.+)\.awsapps\.com\/start|https?:\/\/identitycenter\.amazonaws\.com\/ssoins-[\da-zA-Z]{16})\/?$/
