+++ packages/core/src/login/webview/vue/amazonq/backend_amazonq.ts
@@ -19,9 +19,9 @@ import { VSCODE_EXTENSION_ID } from '../../../../shared/extensions'
 import { getLogger } from '../../../../shared/logger'
 import { debounce } from 'lodash'
 import { AuthError, AuthFlowState, userCancelled } from '../types'
-import { builderIdStartUrl } from '../../../../auth/sso/model'
 import { ToolkitError } from '../../../../shared/errors'
 import { withTelemetryContext } from '../../../../shared/telemetry/util'
+import { builderIdStartUrl } from '../../../../auth/sso/constants'
 
 const className = 'AmazonQLoginWebview'
 export class AmazonQLoginWebview extends CommonAuthWebview {
